Python-Scripts/requirements.txt
netmiko==2.4.2
PyYAML==5.1.2
netdev==0.9.3
requests==2.22.0